E3-ai-shoe-starter-main/requirements.txt
streamlit
google.cloud
google-cloud-bigquery
google-auth
google-auth-oauthlib
streamlit-auth0
google-cloud-storage
python-dotenv
google-cloud-aiplatform
google-generativeai